logaritmos controlados

+++ methods/part1/objects.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 class Bisection(BaseModel):
     fun: str 
-    a: int 
-    b: int
+    a: float 
+    b: float
     tol: float
     niter: int
